video component WIP
