+++ package.json
@@ -6,6 +6,9 @@
     "@testing-library/jest-dom": "^4.2.4",
     "@testing-library/react": "^9.4.0",
     "@testing-library/user-event": "^7.2.1",
+    "mobx": "^5.15.4",
+    "mobx-react": "^6.1.7",
+    "mobx-utils": "^5.5.3",
     "react": "^16.12.0",
     "react-dom": "^16.12.0",
     "react-scripts": "3.3.0"
@@ -32,6 +35,7 @@
     ]
   },
   "devDependencies": {
+    "babel-preset-mobx": "^2.0.0",
     "eslint": "^6.8.0",
     "eslint-config-standard": "^14.1.0",
     "eslint-plugin-import": "^2.20.1",
